Split ARN specific code into a dedicates arn.py
