+++ plugins/module_utils/core.py
@@ -53,7 +53,8 @@ don't need to be wrapped in the backoff decorator.
 from __future__ import (absolute_import, division, print_function)
 __metaclass__ = type
 
-import re
+# Used to live here, moved into ansible_collections.amazon.aws.plugins.module_utils.arn
+from .arn import parse_aws_arn  # pylint: disable=unused-import
 
 # Used to live here, moved into ansible_collections.amazon.aws.plugins.module_utils.botocore
 from .botocore import HAS_BOTO3  # pylint: disable=unused-import
@@ -96,25 +97,3 @@ def scrub_none_parameters(parameters, descend_into_lists=True):
             clean_parameters[k] = v
 
     return clean_parameters
-
-
-def parse_aws_arn(arn):
-    """
-    The following are the general formats for ARNs.
-        arn:partition:service:region:account-id:resource-id
-        arn:partition:service:region:account-id:resource-type/resource-id
-        arn:partition:service:region:account-id:resource-type:resource-id
-    The specific formats depend on the resource.
-    The ARNs for some resources omit the Region, the account ID, or both the Region and the account ID.
-    """
-    m = re.search(r"arn:(aws(-([a-z\-]+))?):(\w+):([a-z0-9\-]*):(\d*):(.*)", arn)
-    if m is None:
-        return None
-    result = dict()
-    result.update(dict(partition=m.group(1)))
-    result.update(dict(service=m.group(4)))
-    result.update(dict(region=m.group(5)))
-    result.update(dict(account_id=m.group(6)))
-    result.update(dict(resource=m.group(7)))
-
-    return result
